add cpp examples to mkdocs as snippets
