+++ .github/workflows/ci.yml
@@ -36,7 +36,7 @@ jobs:
 
   cpp-conan-test:
     needs: [ cpp-build ]
-    name: "Test package with example"
+    name: "Test package with examples"
     runs-on: ubuntu-latest
     defaults:
       run:
@@ -51,14 +51,14 @@ jobs:
       - name: Build test package
         run: CONAN_SOURCE_DIR=`pwd`/.. conan create . drift/stable
 
-      - name: Build example
+      - name: Build examples
         run: |
           mkdir example/build
           cd example/build
           cmake -DCMAKE_BUILD_TYPE=Release ..
           cmake --build . --target all
-      - name: Run example
-        run: ./example/build/bin/drift_protocol_example
+      - name: Run examples
+        run: ./examples/build/bin/drift_protocol_example
 
   cpp-conan-upload:
     needs: [ cpp-conan-test ]
