+++ cpp/examples/drift_package_example.cc
