+++ cpp/examples/payload_with_image_example.cc
@@ -1,20 +1,25 @@
 #include <drift_protocol/common/data_payload.pb.h>
 #include <drift_protocol/common/drift_package.pb.h>
 #include <google/protobuf/util/time_util.h>
+
 #include <wavelet_buffer/wavelet_buffer.h>
+#include <wavelet_buffer/img/wavelet_image.h>
+#include <wavelet_buffer/img/jpeg_codecs.h>
 #include <wavelet_buffer/denoise_algorithms.h>
+
 
 using drift::proto::common::DataPayload;
 using drift::proto::common::DriftPackage;
 using drift::proto::common::StatusCode;
 using drift::proto::meta::MetaInfo;
-using drift::proto::meta::ScalarValuesInfo;
+using drift::proto::meta::ImageInfo;
 
 using drift::WaveletBuffer;
 using drift::WaveletParameters;
 using drift::WaveletTypes;
-using drift::Signal1D;
-using NoDenoise = drift::NullDenoiseAlgorithm<float>;
+using drift::img::WaveletImage;
+using Denoiser = drift::SimpleDenoiseAlgorithm<float>;
+using drift::img::RgbJpegCodec;
 
 using google::protobuf::util::TimeUtil;
 using google::protobuf::Any;
@@ -23,8 +28,21 @@ int main() {
     const auto pb_time = TimeUtil::GetCurrentTime();
     std::string message;
 
-    const Signal1D kData = {0, 10., 0.3, 1};
-    const std::vector<std::string> kNames = {"x0", "x1", "x2", "x3"};
+    // Load and decompose image with wavelet image
+    const auto kWidth = 800;
+    const auto kHeight = 500;
+    WaveletImage img(WaveletParameters{
+            .signal_shape = {kWidth, kHeight},
+            .signal_number = 3,
+            .decomposition_steps = 3,
+            .wavelet_type = WaveletTypes::kDB3,
+    });
+
+    // Load JPG image, decode, decompose with wavelet transformation and denoise
+    if (auto code = img.ImportFromFile(IMAGE_PATH, Denoiser(0.9), RgbJpegCodec())) {
+        std::cerr << "Failed to import image " << code;
+        return -1;
+    }
 
     {
         // Create a package and serialize it
@@ -35,39 +53,27 @@ int main() {
         original.mutable_publish_timestamp()->CopyFrom(pb_time);
 
         // Fill meta data
-        ScalarValuesInfo info;
-        for (const auto &name: kNames) {
-            info.add_variables()->set_name(name);
-        }
+        ImageInfo image_info;
+        image_info.set_type(ImageInfo::WB); // this is image decode with WaveletBuffer
+        image_info.set_width(kWidth);
+        image_info.set_height(kHeight);
+        image_info.set_channel_layout("RGB");
 
         MetaInfo meta;
-        meta.set_type(MetaInfo::SCALAR_VALUES);
-        meta.mutable_scalar_info()->CopyFrom(info);
+        meta.set_type(MetaInfo::IMAGE);
+        meta.mutable_image_info()->CopyFrom(image_info);
 
         original.mutable_meta()->CopyFrom(meta);
 
-        // Decompose and compress signal
-        WaveletBuffer buffer(WaveletParameters{
-                .signal_shape =  {kData.size()},
-                .signal_number = 1,
-                .decomposition_steps = 0,
-                .wavelet_type = WaveletTypes::kNone, // No composition, wavelet buffer is a vector now
-        });
-
-        if (!buffer.Decompose(kData, NoDenoise())) {
-            std::cerr << "Failed decompose the signal";
+        // Prepare payload with the image
+        std::string compressed_image;
+        if (!img.buffer().Serialize(&compressed_image, 16)) {
+            std::cerr << "Failed to compress image image";
             return -1;
         }
 
-        std::string data;
-        if (!buffer.Serialize(&data, 16)) {
-            std::cerr << "Failed decompose the signal";
-            return -1;
-        }
-
-        // Prepare payload
         DataPayload payload;
-        payload.set_data(data);
+        payload.set_data(compressed_image);
         original.add_data()->PackFrom(payload);
 
         // Serialize package to message
