+++ cpp/examples/payload_with_timeseries_example.cc
@@ -1,25 +1,20 @@
 #include <drift_protocol/common/data_payload.pb.h>
 #include <drift_protocol/common/drift_package.pb.h>
 #include <google/protobuf/util/time_util.h>
-
 #include <wavelet_buffer/wavelet_buffer.h>
-#include <wavelet_buffer/img/wavelet_image.h>
-#include <wavelet_buffer/img/jpeg_codecs.h>
 #include <wavelet_buffer/denoise_algorithms.h>
-
 
 using drift::proto::common::DataPayload;
 using drift::proto::common::DriftPackage;
 using drift::proto::common::StatusCode;
 using drift::proto::meta::MetaInfo;
-using drift::proto::meta::ImageInfo;
+using drift::proto::meta::TimeSeriesInfo;
 
 using drift::WaveletBuffer;
 using drift::WaveletParameters;
 using drift::WaveletTypes;
-using drift::img::WaveletImage;
-using NoDenoise = drift::SimpleDenoiseAlgorithm<float>;
-using drift::img::RgbJpegCodec;
+using drift::Signal1D;
+using Denoiser = drift::ThresholdAbsDenoiseAlgorithm<float>;
 
 using google::protobuf::util::TimeUtil;
 using google::protobuf::Any;
@@ -28,22 +23,7 @@ int main() {
     const auto pb_time = TimeUtil::GetCurrentTime();
     std::string message;
 
-    // Load and decompose image with wavelet image
-    const auto kWidth = 800;
-    const auto kHeight = 500;
-    WaveletImage img(WaveletParameters{
-            .signal_shape = {kWidth, kHeight},
-            .signal_number = 3,
-            .decomposition_steps = 3,
-            .wavelet_type = WaveletTypes::kDB3,
-    });
-
-    // Load JPG image, decode, decompose with wavelet transformation and denoise
-    if (auto code = img.ImportFromFile(IMAGE_PATH, NoDenoise(0.9), RgbJpegCodec())) {
-        std::cerr << "Failed to import image " << code;
-        return -1;
-    }
-
+    const Signal1D kTimeSeries{0.1, 0.2, 0.5, 0.1, 0.2, 0.1, 0.6, 0.1, 0.1, 0.2};
     {
         // Create a package and serialize it
         DriftPackage original;
@@ -53,27 +33,44 @@ int main() {
         original.mutable_publish_timestamp()->CopyFrom(pb_time);
 
         // Fill meta data
-        ImageInfo image_info;
-        image_info.set_type(ImageInfo::WB); // this is image decode with WaveletBuffer
-        image_info.set_width(kWidth);
-        image_info.set_height(kHeight);
-        image_info.set_channel_layout("RGB");
+        TimeSeriesInfo info;
+        info.mutable_start_timestamp()->CopyFrom(pb_time - TimeUtil::SecondsToDuration(1));
+        info.mutable_stop_timestamp()->CopyFrom(pb_time);
+        info.set_size(kTimeSeries.size());
+        info.set_first(kTimeSeries[0]);
+        info.set_last(kTimeSeries[kTimeSeries.size() - 1]);
+        info.set_min(blaze::min(kTimeSeries));
+        info.set_max(blaze::max(kTimeSeries));
+        info.set_min(blaze::min(kTimeSeries));
 
         MetaInfo meta;
-        meta.set_type(MetaInfo::IMAGE);
-        meta.mutable_image_info()->CopyFrom(image_info);
+        meta.set_type(MetaInfo::TIME_SERIES);
+        meta.mutable_time_series_info()->CopyFrom(info);
 
         original.mutable_meta()->CopyFrom(meta);
 
-        // Prepare payload with the image
-        std::string compressed_image;
-        if (!img.buffer().Serialize(&compressed_image, 16)) {
-            std::cerr << "Failed to compress image image";
+        // Decompose and compress signal
+        WaveletBuffer buffer(WaveletParameters{
+                .signal_shape =  {kTimeSeries.size()},
+                .signal_number = 1,
+                .decomposition_steps = 2,
+                .wavelet_type = WaveletTypes::kDB1, // Haar
+        });
+
+        if (!buffer.Decompose(kTimeSeries, Denoiser(0, 0.1))) {
+            std::cerr << "Failed decompose the signal";
             return -1;
         }
 
+        std::string data;
+        if (!buffer.Serialize(&data, 16)) {
+            std::cerr << "Failed decompose the signal";
+            return -1;
+        }
+
+        // Prepare payload
         DataPayload payload;
-        payload.set_data(compressed_image);
+        payload.set_data(data);
         original.add_data()->PackFrom(payload);
 
         // Serialize package to message
