+++ cpp/examples/trigger_message_example.cc
