+++ docs/cpp_examples/drift_package_example.md
@@ -1,0 +1,7 @@
+# Basic Drift Package Example
+
+This example shows how to serialize and parse a [Drift Package](/docs/api/common) in C++.
+
+```cpp title="cpp/examples/drift_package_example.cc"
+--8<-- "../cpp/examples/drift_package_example.cc:"
+```
