+++ docs/cpp_examples/payload_with_image_example.md
@@ -1,0 +1,9 @@
+# Drift Package With Image
+
+Here you can see how to load a JPEG image, compress it
+with [WaveletBuffer](https://github.com/panda-official/WaveletBuffer) and build a [Drift Package](/docs/api/common) with
+[meta information](/docs/api/meta) in C++.
+
+```cpp title="cpp/examples/payload_with_image_example.cc"
+--8<-- "../cpp/examples/payload_with_image_example.cc"
+```
