+++ docs/cpp_examples/payload_with_scalars_example.md
@@ -1,0 +1,8 @@
+# Drift Package With Scalar Values
+
+Here you can see how to compress scalar values data with [WaveletBuffer](https://github.com/panda-official/WaveletBuffer)
+and build a [Drift Package](/docs/api/common) with [meta information](/docs/api/meta) in C++.
+
+```cpp title="cpp/examples/payload_with_scalars.cc"
+--8<-- "../cpp/examples/payload_with_scalars.cc"
+```
