+++ docs/cpp_examples/payload_with_timeseries_example.md
@@ -1,0 +1,8 @@
+# Drift Package With Time Series
+
+Here you can see how to compress time series data with [WaveletBuffer](https://github.com/panda-official/WaveletBuffer)
+and build a [Drift Package](/docs/api/common) with [meta information](/docs/api/meta) in C++.
+
+```cpp title="cpp/examples/payload_with_timeseries_example.cc"
+--8<-- "../cpp/examples/payload_with_timeseries_example.cc"
+```
