+++ docs/cpp_examples/trigger_message_example.md
@@ -1,0 +1,8 @@
+# Drift Package With Trigger
+
+Here you can see how to build a [Drift Package](/docs/api/common) with [a trigger message](/docs/api/triggering) for a
+time interval in C++.
+
+```cpp title="cpp/examples/trigger_message_example.cc"
+--8<-- "../cpp/examples/trigger_message_example.cc"
+```
