--- docs/use_cases.md
@@ -1,28 +1,0 @@
-# Typical Use Cases for C++
-
-
-## Drift Package
-
-T
-
-## Serialize / Deserialize a Trigger Message
-
-
-
-## Serialize / Deserialize an Image
-
-Python
-
-C++
-
-## Serialize / Deserialize a Time Series
-
-Python
-
-C++
-
-## Serialize / Deserialize Scalar Values
-
-Python
-
-C++
